chore: rename identification utils file
