+++ src/index.ts
@@ -1,4 +1,4 @@
-import type { AuthorizationContext } from "./utils/identification";
+import type { AuthorizationContext } from "./utils/auth";
 import { fixBase64EncodingOnKey } from "./utils/jwk";
 // polyfill due to known bugs on URL implementation for react native
 // https://github.com/facebook/react-native/issues/24428
